models folder renamed to schemas
